#2732 improve engine monitoring ingestion
