+++ discovery-frontend/src/assets/css/theme_dark/metatron/page.css
@@ -42445,7 +42445,6 @@
     margin-left: 5px;
     width: auto;
     min-width: 320px;
-    max-width: 520px;
     z-index: 10;
 }
 
@@ -43033,6 +43032,8 @@
     color: #4e5368;
     font-size: 13px;
 }
+.theme-dark table.ddp-table-detail3 tr:last-of-type td
+{border-bottom:none;}
 
 .theme-dark table.ddp-table-detail3 tr:nth-child(odd) td {
     background-color: #fafafa;
